isejda

+++ backend/actionTable.php
@@ -1,13 +1,7 @@
 <?php
 session_start();
 require_once(__DIR__ . '/../include/conn.php');
-// insert record
 global $conn;
-
-//if (isset($_SESSION['role'])) {
-//print_r($_SESSION['role']);
-//exit;
-//}
 
 if(isset($_POST['operation'])){
     if($_POST['operation'] == "Insert") {
@@ -16,8 +10,6 @@ if(isset($_POST['operation'])){
         $request = $_REQUEST;
         $userRole = $_SESSION['role'];
         $userId = $_SESSION['id'];
-//        print_r($userId);
-//        print_r($userRole);
 
         $sql= "";
 //create col like table in database
@@ -32,30 +24,48 @@ if(isset($_POST['operation'])){
 
         if ($userRole == 'admin') {
             // Fetch data for admin (all users)
-            $sql = "SELECT * FROM users WHERE role = 'user'";
-        }elseif ($userRole == 'user') {
+            $sql = "SELECT * FROM users WHERE role = 'user' || role = 'manager'";
+        }elseif ($userRole == 'user' || $userRole == 'manager') {
             // Fetch data for manager (users reporting to the manager)
-            $sql = "SELECT u.*
-                    FROM users u
-                    INNER JOIN hierarchy h ON u.id = h.subordinate_id
-                    WHERE h.supervisor_id = $userId OR u.id = $userId";
-        }
+            $sql = "WITH RECURSIVE SupervisedUsers AS (
+                SELECT u.id, u.name, u.lastname, u.email, u.birthday, u.role
+                FROM users u
+                JOIN hierarchy h ON u.id = h.subordinate_id
+                WHERE h.supervisor_id = (SELECT id FROM users WHERE id = $userId)
 
+                UNION ALL
+
+                SELECT u.id, u.name, u.lastname, u.email, u.birthday, u.role
+                FROM users u
+                JOIN hierarchy h ON u.id = h.subordinate_id
+                JOIN SupervisedUsers su ON h.supervisor_id = su.id
+            )
+            SELECT * FROM SupervisedUsers;";
+        }
 
         $query = mysqli_query($conn, $sql);
         $totalData = mysqli_num_rows($query);
         $totalFilter = $totalData;
 
-        if ($userRole == 'admin') {
+/*        if ($userRole == 'admin') {
             // Fetch data for admin (all users)
             $sql = "SELECT * FROM users WHERE role = 'user'";
-        }elseif ($userRole == 'user') {
+        }elseif ($userRole == 'user' || $userRole == 'manager') {
             // Fetch data for manager (users reporting to the manager)
+            $sql = "WITH RECURSIVE SupervisedUsers AS (
+                SELECT u.id, u.name, u.lastname, u.email, u.birthday, u.role
+                FROM users u
+                JOIN hierarchy h ON u.id = h.subordinate_id
+                WHERE h.supervisor_id = (SELECT id FROM users WHERE id = $userId)
 
-            $sql = "SELECT u.*
-                    FROM users u
-                    INNER JOIN hierarchy h ON u.id = h.subordinate_id
-                    WHERE h.supervisor_id = $userId OR u.id = $userId";
+                UNION ALL
+
+                SELECT u.id, u.name, u.lastname, u.email, u.birthday, u.role
+                FROM users u
+                JOIN hierarchy h ON u.id = h.subordinate_id
+                JOIN SupervisedUsers su ON h.supervisor_id = su.id
+            )
+            SELECT * FROM SupervisedUsers;";
         }
 
         //searchi funksional
@@ -67,7 +77,61 @@ if(isset($_POST['operation'])){
             $sql .= " OR role Like '%" . $request['search']['value'] . "%' ";
             $sql .= " OR name Like '%" . $request['search']['value'] . "%' )";
 
+        }*/
+
+        if ($userRole == 'admin') {
+            // Fetch data for admin (all users)
+            $sql = "SELECT * FROM users WHERE role = 'user' || role = 'manager'";
+            if (!empty($request['search']['value'])) {
+                $sql .= " AND (id Like '%" . $request['search']['value'] . "%' ";
+                $sql .= " OR lastname Like '%" . $request['search']['value'] . "%' ";
+                $sql .= " OR email Like '%" . $request['search']['value'] . "%' ";
+                $sql .= " OR birthday Like '%" . $request['search']['value'] . "%' ";
+                $sql .= " OR role Like '%" . $request['search']['value'] . "%' ";
+                $sql .= " OR name Like '%" . $request['search']['value'] . "%' )";
+            }
+        }elseif ($totalFilter == 0){
+            $sql = "SELECT * FROM users WHERE id = $userId";
+        } elseif ($userRole == 'user' || $userRole == 'manager') {
+            // Fetch data for manager (users reporting to the manager)
+            $recursiveSql = "
+        WITH RECURSIVE SupervisedUsers AS (
+            SELECT u.id, u.name, u.lastname, u.email, u.birthday, u.role
+            FROM users u
+            JOIN hierarchy h ON u.id = h.subordinate_id
+            WHERE h.supervisor_id = (SELECT id FROM users WHERE id = $userId)
+            
+            UNION ALL
+            
+            SELECT u.id, u.name, u.lastname, u.email, u.birthday, u.role
+            FROM users u
+            JOIN hierarchy h ON u.id = h.subordinate_id
+            JOIN SupervisedUsers su ON h.supervisor_id = su.id
+        )
+        SELECT * FROM SupervisedUsers";
+
+            $sql = "SELECT * FROM ($recursiveSql) AS subquery";
+
+            if (!empty($request['search']['value'])) {
+                $searchValue = mysqli_real_escape_string($conn, $request['search']['value']);
+                $searchConditions = [
+                    "id LIKE '%$searchValue%'",
+                    "lastname LIKE '%$searchValue%'",
+                    "email LIKE '%$searchValue%'",
+                    "birthday LIKE '%$searchValue%'",
+                    "role LIKE '%$searchValue%'",
+                    "name LIKE '%$searchValue%'"
+                ];
+
+                // Add WHERE clause only if there are search conditions
+                $sql .= " WHERE " . implode(" OR ", $searchConditions);
+            }
         }
+
+// Add search functionality
+
+
+
         $query = mysqli_query($conn, $sql);
         $totalData = mysqli_num_rows($query);
 
@@ -104,21 +168,25 @@ if(isset($_POST['operation'])){
                 $subdata[] = $row[3]; //email
                 $subdata[] = $row[4]; //birthday
                 $subdata[] = $row[5]; //role
-                $subdata[] = '<button type="button" name="update" id="' . $row[0] . '" class="btn btn-w-m btn-primary btn-xm update-user-btn"><i class="fa fa-edit">&nbsp;</i>Edit</button>
-                  <button type="button" name="delete" id="' . $row[0] . '" class="btn btn-w-m btn-danger btn-xm delete-user-btn" ><i class="fa fa-trash">&nbsp;</i>Delete</button>';
+                $subdata[] = '
+                            <div class="btn-group">
+                            <button data-toggle="dropdown" class="btn btn-primary dropdown-toggle">Action <span class="caret"></span></button>
+                            <ul class="dropdown-menu">
+                                <li>
+                                <button type="button" name="update" id="' . $row[0] . '" class="btn btn-w-m btn-xm update-user-btn"><i class="fa fa-edit">&nbsp;</i>Edit</button>
+                                </li>
+                                <li class="divider"></li>
+                                <li>
+                                <button type="button" name="delete" id="' . $row[0] . '" class="btn btn-w-m btn-xm delete-user-btn" ><i class="fa fa-trash">&nbsp;</i>Delete</button>
+                                </li>
+                            </ul>
+                            </div>
+                
+               ';
                 /*                  <a href="backend/delete.php?deleteid='.$row[0].'" onclick="return confirm(\'Are you sure?\')" class="btn btn-w-m btn-danger btn-xm"><i class="fa fa-trash">&nbsp;</i>Delete</a>';*/
-
-                if ($row[0] == $userId && !$loggedInUserAdded) {
-                    $subdata = array_map(function($item) {
-                        return '<em><strong>' . $item . '</strong></em>';
-                        }, $subdata);
-                    $loggedInUserAdded = true;
-                    array_unshift($data, $subdata);
-                } else {
                     $data[] = $subdata;
-                }
             }
-        }elseif ($userRole == 'user') {
+        }elseif ($userRole == 'user' || $userRole == 'manager') {
             // Fetch data for manager (users reporting to the manager)
 
             while ($row = mysqli_fetch_array($query)) {
@@ -129,16 +197,7 @@ if(isset($_POST['operation'])){
                 $subdata[] = $row[3]; //email
                 $subdata[] = $row[4]; //birthday
                 $subdata[] = $row[5]; //role
-
-                if ($row[0] == $userId && !$loggedInUserAdded) {
-                    $subdata = array_map(function($item) {
-                        return '<em><strong>' . $item . '</strong></em>';
-                    }, $subdata);
-                    $loggedInUserAdded = true;
-                    array_unshift($data, $subdata);
-                } else {
-                    $data[] = $subdata;
-                }
+                $data[] = $subdata;
             }
 
         }
@@ -154,8 +213,8 @@ if(isset($_POST['operation'])){
         echo json_encode($json_data);
     }
 
-
     if($_POST['operation'] == "Save"){
+        $supervisor_id = $_SESSION['id'];
         $name = $_POST['name'];
         $lastname = $_POST['lastname'];
         $email = $_POST['email'];
@@ -176,13 +235,23 @@ if(isset($_POST['operation'])){
 
             $result = mysqli_query($conn, $query);
             if($result){
-                echo 'Your record has been saved';
+                $subordinate_id = mysqli_insert_id($conn);
+                $hierarchy_query = "INSERT INTO hierarchy (supervisor_id, subordinate_id)
+                                VALUES ('$supervisor_id', '$subordinate_id')";
+                $hierarchy_result = mysqli_query($conn, $hierarchy_query);
+
+                if($hierarchy_result){
+                    echo 'Your record has been saved. ';
+                } else {
+                    echo "Failed to save hierarchy information.";
+                }
             }
             else{
                 echo "Your record has not been saved";
             }
         }
     }
+
     if($_POST['operation'] == "Edit"){
         if (isset($_POST['member_id'])) {
             $id = $_POST['member_id'];
@@ -191,6 +260,8 @@ if(isset($_POST['operation'])){
             $email = $_POST['email'];
             $birthday = $_POST['birthday'];
             $selectedRole = $_POST['role'];
+/*            print_r($selectedRole);
+            exit;*/
             $password = $_POST['password'];
             $hashed_password = password_hash($password, PASSWORD_DEFAULT);
             $select = "select * from users where id='$id'";
@@ -241,12 +312,22 @@ if(isset($_POST['operation'])){
     if($_POST['operation'] == "Delete") {
         if (isset($_POST['member_id'])) {
             $id = $_POST['member_id'];
-            $sql = " DELETE FROM users
+            $query = "DELETE FROM hierarchy 
+            WHERE subordinate_id = $id OR supervisor_id = $id";
+
+            $res = $conn->query($query);
+
+            if($res){
+                $sql = " DELETE FROM users
             WHERE id = $id";
 
-            $result = $conn->query($sql);
-            if($result){
-                echo "Delete is successful";
+                $result = $conn->query($sql);
+                if($result){
+                    echo "Delete is successful";
+                }
+                else{
+                    die($conn->error);
+                }
             }
             else{
                 die($conn->error);
